+++ test/ship-mate-test.el
@@ -774,139 +774,277 @@
 (ert-deftest ship-mate-submarine--recompile ()
   :tags '(submarine)
 
-  (let ((ship-mate-submarine--in-progress t))
+  (let ((ship-mate--last-compilation-type 'other)
+        (ship-mate-submarine--processes '(process)))
 
     (should-error (ship-mate-submarine--recompile)))
 
-  (let ((ship-mate--last-compilation-type 'other))
+  (let ((ship-mate--last-compilation-type nil)
+        (ship-mate-submarine--processes '(process)))
 
     (should-error (ship-mate-submarine--recompile)))
 
-  (let ((ship-mate--last-compilation-type nil))
+  (let ((ship-mate--last-compilation-type 'ship-mate)
+        (ship-mate--last-command 'test))
 
-    (should-error (ship-mate-submarine--recompile)))
+    (bydi (ship-mate-submarine--run)
 
-  (ert-with-test-buffer (:name "open window")
+      (ship-mate-submarine--recompile)
 
-    (pop-to-buffer (current-buffer))
+      (bydi-was-called ship-mate-submarine--run))))
 
-    (setq-local ship-mate--this-command 'test)
+(ert-deftest ship-mate-submarine--in-progress ()
+  :tags '(submarine)
 
-    (bydi (quit-window
-           recompile)
+  (let ((ship-mate-submarine--processes nil))
 
-      (shut-up (ship-mate-submarine--recompile))
+    (should-not (ship-mate-submarine--in-progress))
 
-      (bydi-was-called quit-window)))
+    (push 'process ship-mate-submarine--processes)
 
-  (let ((ship-mate-submarine--in-progress nil)
-        (ship-mate-submarine--buffer nil))
-    (bydi ((:watch ship-mate-submarine--in-progress)
-           (:watch ship-mate-submarine--buffer)
-           (:watch display-buffer-alist)
-           recompile)
+    (should (ship-mate-submarine--in-progress))))
 
-      (shut-up (ship-mate-submarine--recompile))
+(ert-deftest ship-mate-submarine--run ()
+  :tags '(submarine)
 
-      (bydi-was-called recompile)
-      (bydi-was-set display-buffer-alist)
-      (bydi-was-set ship-mate-submarine--in-progress)
-      (bydi-was-set ship-mate-submarine--buffer))))
+  (bydi ((:ignore ship-mate-submarine--ensure-no-hidden-buffers)
+         (:watch display-buffer-alist))
+
+    (ert-with-test-buffer (:name "sub-run")
+
+      (let ((ship-mate--current-command-name "test"))
+        (shut-up (ship-mate-submarine--run (lambda () (current-buffer)))))
+
+      (should (buffer-local-value 'ship-mate--hidden (current-buffer)))
+      (bydi-was-called ship-mate-submarine--ensure-no-hidden-buffers)
+      (bydi-was-set display-buffer-alist))))
+
+(ert-deftest ship-mate-submarine--hide ()
+  :tags '(submarine)
+
+  (let ((ship-mate--this-command 'test))
+
+    (bydi ((:othertimes ship-mate-submarine--get-process)
+           (:ignore ship-mate-submarine--watch-process)
+           (:watch ship-mate--hidden)
+           quit-window)
+
+      (should-error (ship-mate-submarine--hide))
+
+      (bydi-toggle-sometimes)
+
+      (ert-with-test-buffer (:name "sub-hide")
+        (shut-up (ship-mate-submarine--hide))
+
+        (bydi-was-called ship-mate-submarine--watch-process)
+        (bydi-was-set-to ship-mate--hidden t)))))
+
+(ert-deftest ship-mate-submarine--get-process ()
+  :tags '(submarine)
+
+  (should-not (ship-mate-submarine--get-process nil))
+
+  (ert-with-test-buffer (:name "sub-proc")
+
+    (should-not (ship-mate-submarine--get-process (current-buffer)))
+
+    (bydi ((:spy get-buffer-process)
+           (:always ship-mate--command-buffer-p))
+
+      (ship-mate-submarine--get-process (current-buffer))
+
+      (bydi-was-called get-buffer-process))))
+
+(ert-deftest ship-mate-submarine--ensure-no-hidden-buffer ()
+  :tags '(submarine)
+
+  (ert-with-test-buffer (:name "sub-ensure")
+
+    (setq ship-mate--hidden t)
+
+    (bydi ((:spy quit-window))
+
+      (pop-to-buffer (current-buffer))
+
+      (ship-mate-submarine--ensure-no-hidden-buffers)
+
+      (bydi-was-called quit-window))))
 
 (ert-deftest ship-mate-submarine--check ()
   :tags '(submarine)
 
-  (let ((ship-mate-submarine--process nil))
-
-    (should-error (ship-mate-submarine--check)))
-
-  (let ((ship-mate-submarine--process 'process)
-        (ship-mate-submarine--timer 'timer)
-        (ship-mate-submarine--in-progress t)
-        (ship-mate-prompt-for-hidden-buffer t))
-
-    (bydi ((:sometimes process-live-p)
-           (:mock process-exit-status :return 0)
-           cancel-timer
-           ship-mate-submarine--surface
-           (:watch ship-mate-submarine--in-progress)
-           run-with-idle-timer)
+  (let ((ship-mate-submarine--processes nil))
+    (bydi ((:ignore ship-mate-submarine--clear-timer)
+           (:spy ship-mate-submarine--surface))
 
       (ship-mate-submarine--check)
 
-      (bydi-was-not-called ship-mate-submarine--surface)
+      (bydi-was-called ship-mate-submarine--clear-timer)
+      (bydi-was-not-called ship-mate-submarine--surface)))
 
-      (bydi-toggle-sometimes)
+  (let ((ship-mate-submarine--processes '(a b))
+        (ship-mate-prompt-for-hidden-buffer nil))
 
-      (ship-mate-submarine--check)
+    (ert-with-test-buffer (:name "sub-check")
+      (bydi ((:ignore ship-mate-submarine--clear-timer)
+             (:sometimes process-live-p)
+             (:mock process-buffer :return (current-buffer))
+             (:mock process-exit-status :return 'status)
+             (:sometimes ship-mate-submarine--hidden-buffer-p)
+             run-with-idle-timer
+             ship-mate-submarine--clear-process
+             ship-mate-submarine--surface)
 
-      (bydi-was-set ship-mate-submarine--in-progress)
-      (bydi-was-called run-with-idle-timer)
+        ;; Process is live, won't surface.
+        (ship-mate-submarine--check)
+        (bydi-was-not-called ship-mate-submarine--surface)
 
-      (setq ship-mate-submarine--process 'process
-            ship-mate-submarine--timer 'timer
-            ship-mate-submarine--in-progress t
-            ship-mate-prompt-for-hidden-buffer nil)
+        (bydi-toggle-volatile 'process-live-p)
 
-      (ship-mate-submarine--check)
+        ;; Process is dead, will surface.
+        (ship-mate-submarine--check)
+        (bydi-was-called ship-mate-submarine--surface t)
 
-      (bydi-was-called ship-mate-submarine--surface))))
+        ;; Runs with idle timer if set.
+        (setq ship-mate-prompt-for-hidden-buffer t)
+        (ship-mate-submarine--check)
+        (bydi-was-called ship-mate-submarine--clear-process t)
+        (bydi-was-not-called ship-mate-submarine--surface)
+        (bydi-was-called run-with-idle-timer t)
+
+        ;; Clears process if not a hidden buffer.
+        (bydi-toggle-volatile 'ship-mate-submarine--hidden-buffer-p)
+        (ship-mate-submarine--check)
+        (bydi-was-called ship-mate-submarine--clear-process)
+        (bydi-was-not-called ship-mate-submarine--surface)))))
+
+(ert-deftest ship-mate-submarine--hidden-buffer-p ()
+  :tags '(submarine)
+
+  (ert-with-test-buffer (:name "sub-hidden-p")
+    (should-not (ship-mate-submarine--hidden-buffer-p (current-buffer)))
+
+    (setq ship-mate--hidden t)
+
+    (should (ship-mate-submarine--hidden-buffer-p (current-buffer)))))
 
 (ert-deftest ship-mate-submarine--delayed-prompt ()
+  :tags '(submarine)
+
   (bydi ((:always yes-or-no-p)
          ship-mate-submarine--surface)
 
-    (ship-mate-submarine--delayed-prompt (current-time) 0)
-    (ship-mate-submarine--delayed-prompt (current-time) 1)
+    (ship-mate-submarine--delayed-prompt (current-time) 0 (current-buffer))
+    (ship-mate-submarine--delayed-prompt (current-time) 1 (current-buffer))
 
     (bydi-was-called-n-times ship-mate-submarine--surface 2)))
 
 (ert-deftest ship-mate-submarine--watch-process ()
   :tags '(submarine)
 
-  (let ((ship-mate-submarine--in-progress nil)
-        (ship-mate-submarine--process nil)
+  (let ((ship-mate-submarine--processes nil)
         (ship-mate-submarine--timer))
 
     (bydi (run-with-timer
-           (:watch ship-mate-submarine--timer))
+           (:mock process-buffer :return (current-buffer))
+           (:othertimes ship-mate--command-buffer-p)
+           (:watch ship-mate-submarine--timer)
+           (:watch ship-mate-submarine--processes))
 
       (ship-mate-submarine--watch-process 'process)
 
       (bydi-was-not-set ship-mate-submarine--timer)
 
-      (setq ship-mate-submarine--in-progress t)
+      (bydi-toggle-sometimes)
 
       (ship-mate-submarine--watch-process 'process)
 
+      (bydi-was-called run-with-timer)
+      (bydi-was-set ship-mate-submarine--timer)
+      (bydi-was-set ship-mate-submarine--processes))))
+
+(ert-deftest ship-mate-submarine--clear-process ()
+  :tags '(submarine)
+
+  (let ((ship-mate-submarine--processes '(a c)))
+
+    (bydi (ship-mate-submarine--clear-timer
+           (:watch ship-mate-submarine--processes))
+
+      (ship-mate-submarine--clear-process 'b)
+
+      ;; Always clear timer, don't update processes for unknown
+      ;; processes.
+      (bydi-was-called ship-mate-submarine--clear-timer)
+      (bydi-was-not-set ship-mate-submarine--processes)
+
+      ;; Clear process.
+      (ship-mate-submarine--clear-process 'a)
+      (bydi-was-set ship-mate-submarine--processes))))
+
+(ert-deftest ship-mate-submarine--clear-timer ()
+  :tags '(submarine)
+
+  (let ((ship-mate-submarine--timer 'timer))
+
+    (bydi ((:sometimes ship-mate-submarine--in-progress)
+           cancel-timer
+           (:watch ship-mate-submarine--timer))
+
+      ;; Doesn't clear if in progress.
+      (ship-mate-submarine--clear-timer)
+      (bydi-was-not-called cancel-timer)
+      (bydi-was-not-set ship-mate-submarine--timer)
+
+      (bydi-toggle-sometimes)
+
+      (ship-mate-submarine--clear-timer)
+      (bydi-was-called cancel-timer)
       (bydi-was-set ship-mate-submarine--timer))))
 
 (ert-deftest ship-mate-submarine--surface ()
-  :tags '(submarine)
+  :tags '(submarine user-facing)
 
-  (let ((ship-mate-submarine--buffer nil))
+  (should-error (ship-mate-submarine--surface nil))
 
-    (bydi (pop-to-buffer
-           ship-mate-submarine--clear
-           (:watch ship-mate-submarine--buffer)
-           (:ignore ship-mate-submarine--ensure-no-ship-mate-buffers))
+  (bydi (ship-mate-submarine--clear-process
+         (:othertimes ship-mate--buffer-visible-p)
+         (:watch ship-mate--hidden)
+         (:spy pop-to-buffer))
 
-      (should-error (ship-mate-submarine--surface))
+    (ert-with-test-buffer (:name "sub-surf")
 
-      (setq ship-mate-submarine--buffer 'buffer)
+      (ship-mate-submarine--surface (current-buffer))
 
-      (ship-mate-submarine--surface)
+      (bydi-was-called ship-mate-submarine--clear-process)
+      (bydi-was-set-to ship-mate--hidden nil)
+      (bydi-was-called pop-to-buffer t)
 
-      (bydi-was-set ship-mate-submarine--buffer)
-      (bydi-was-called ship-mate-submarine--clear)
-      (bydi-was-called pop-to-buffer))))
+      ;; Doesn't surface already visible buffer.
+      (bydi-toggle-sometimes)
+      (ship-mate-submarine--surface (current-buffer))
+      (bydi-was-not-called pop-to-buffer))))
 
 (ert-deftest ship-mate-show-hidden ()
-  (bydi (ship-mate-submarine--surface)
+  :tags '(user-facing submarine)
 
-    (ship-mate-show-hidden)
+  (bydi (ship-mate-submarine--surface
+         (:mock process-buffer :return 'buffer)
+         (:mock ship-mate--complete-buffer :with (lambda (_p l) (funcall l '(a . buffer)))))
 
-    (bydi-was-called ship-mate-submarine--surface)))
+    (let ((ship-mate-submarine--processes '(process)))
+
+      (call-interactively 'ship-mate-show-hidden)
+
+      (bydi-was-called ship-mate-submarine--surface 'buffer)
+      (bydi-was-not-called ship-mate--complete-buffer)
+
+      (setq ship-mate-submarine--processes '(a b))
+
+      (call-interactively 'ship-mate-show-hidden)
+
+      (bydi-was-called ship-mate-submarine--surface 'buffer)
+      (bydi-was-called ship-mate--complete-buffer))))
 
 (ert-deftest ship-mate-hidden-recompile ()
   :tags '(user-facing submarine)
@@ -915,6 +1053,15 @@
     (ship-mate-hidden-recompile)
 
     (bydi-was-called ship-mate-submarine--recompile)))
+
+(ert-deftest ship-mate-hide ()
+  :tags '(user-facing submarine)
+
+  (bydi ship-mate-submarine--hide
+    (ship-mate-hide)
+
+    (bydi-was-called ship-mate-submarine--hide)))
+
 
 (ert-deftest ship-mate--complete-buffer ()
   :tags '(completion)
@@ -957,6 +1104,14 @@
 
     (bydi-was-called-with pop-to-buffer 'buffer)))
 
+;;; Utiltiy
+
+(ert-deftest ship-mate--buffer-visible-p ()
+  (ert-with-test-buffer (:name "visible")
+    (pop-to-buffer (current-buffer))
+
+    (should (ship-mate--buffer-visible-p (current-buffer)))))
+
 ;;;; Lighter
 
 (ert-deftest ship-mate-mode-lighter--function-segments ()
@@ -978,47 +1133,6 @@
     (bydi (popup-menu)
       (ship-mate-mode-lighter--menu)
       (bydi-was-called popup-menu))))
-
-;;;; Hiding running compilations
-
-(ert-deftest ship-mate-hide ()
-  :tags '(user-facing submarine)
-
-  (bydi ship-mate-submarine--hide
-    (ship-mate-hide)
-
-    (bydi-was-called ship-mate-submarine--hide)))
-
-(ert-deftest ship-mate-submarine--hide ()
-  :tags '(submarine)
-
-  (shut-up
-    (let ((process nil))
-
-      (bydi ((:mock get-buffer-process :return process)
-             (:othertimes ship-mate--command-buffer-p)
-             (:mock ship-mate-submarine--watch-process :return 'hello)
-             (:watch ship-mate-submarine--buffer)
-             (:watch ship-mate-submarine--in-progress)
-             quit-window)
-
-
-        (should-error (ship-mate-submarine--hide))
-
-        (bydi-toggle-sometimes)
-
-        (should-error (ship-mate-submarine--hide))
-
-        (setq process 'process)
-
-        (ship-mate-submarine--hide)
-
-        (bydi-was-set ship-mate-submarine--buffer)
-        (bydi-was-set ship-mate-submarine--in-progress)
-
-        (bydi-was-called-with ship-mate-submarine--watch-process 'process))
-
-      (setq ship-mate-submarine--in-progress nil))))
 
 ;;; ship-mate-test.el ends here